Support raw schemas in addition to Swagger/OpenAPI documents
This adds support for supplying raw schemas instead of Swagger/OpenAPI documents,
using an option --raw-schema in combination with --version.

This enables producing types for stand-alone schemas, which is useful for large
APIs where schemas have been extracted to separate files, and where there may
be no components object in the document.

For V3, the output differs from the output produced for OpenAPI3 documents in that
`schemas: { ... }` is the exported interface instead of
`components: { schemas: { ... } }`.

For V2, `definitions: { ... }` is still exported.