Fix requestBodies that use hypenated naming
If a request body's name is hypenated it breaks when prettier formats the code as it is an invalid object.

Wrapping the request body's name in quotes resolves the issue, and prettier will strip the quotes when they're not required.

Example for reproducing:

```
npx openapi-typescript https://raw.githubusercontent.com/ashsmith/api-specs/fix-invalid-examples/reference/channels.v3.yml --output channels.v3.ts
```