Change from using "any" to "unknown" for unspecified types.
In accordance with Typescript guides the use of the "any" type doesn't
enforce many guarantees, it will match any type.  In situations where
the type is unspecified it is better to use the "unknown" which is
closely related to the "any" type but requires a type assertion or
control flow based narrowing before the type can be used.  The use of
the "unknown" type seems much more appropriate to use then "any" for
these cases.

See:

https://devblogs.microsoft.com/typescript/announcing-typescript-3-0-rc-2/#the-unknown-type

For more details.