Use overridable fetch instead of unidici's request
When loading the schema, the `load` function used to only consider the
Content-Type header when capitalized. However, the HTTP specification
states that headers are to be treated as case-insensitive. HTTP/2
further enforces this by stating that all HTTP headers must be converted
to lowercase when encoded. Therefore, many recent implementations move
away from the conventional capitalization in favour of headers in
lowercase, such as the Nest framework.

In issue #998, it is shown that loading the schema from a live server
implemented in Nest fails when the endpoint does not include the desired
extension.

This commit fixes the issue by replacing unidici's request by fetch. The
request function returns headers as a record, where the keys are left
exactly as the headers are received, meaning that the casing of headers
affect the output. Fetch, on the other hand, returns header in a special
object whose accessor method `get` will lookup headers in a
case-insensitive way.

We took this opportunity to make the fetch implementation overridable.
This way, it is possible for the caller to swap it out for another. It
is step towards portability, as openapiTS does not run as-is in the
browser, Deno or Cloudflare Workers at the moment.

Additionally, unit tests have been added for the `load` function.

Fixes #998.