Make params non-nullable even if all params are optional
In 5.x, an optional query parameter's type could be referenced with an
expression such as

    type myQueryType = operations['myOperation']['parameters']['query']

In 6.0, that only works if there is at least one mandatory query
parameter in addition to the optional one. This makes query parameter
types much more awkward to reference in application code if all the
parameters are optional.

Revert to the 5.x behavior: if there are any parameters, optional or
not, the `parameters` object is always present. If there are any query
parameters, the `query` object is always present, same with `path` and
the other parameter types.

Fixes #995