Drop init argument when not needed in query key
For urls that do not include an init param, generate a query key of length 2. This allows the value to be passed directly to `invalidateQueries()`.

Without this change, the package generates `["get", "/foo", undefined]` which does not correctly match the corresponding get query.

Related to https://github.com/openapi-ts/openapi-typescript/issues/1806