feat: Attach fetch response to SWR data/error
It is useful to have the original fetch response available in the hook result. For instance to distinguish the HTTP status code in case of an `error`.
Since SWR only returns `error` / `data` properties, this provides an escape hatch to be able to access to original `fetch` response, as returned by the `openapi-fetch` client.