docs(openapi-fetch): remove line about new examples (#2339)
As discussed in the maintainer's chat: It seems that examples mostly
showcase how to integrate an (arbitrary) function/client into a
specific framework.

As a result, the examples have very little openapi-fetch specific
content, but a lot of framework specific content. This, in our
opinion, tips the balance of whether the maintenance burden is worth
the ease of use.

Co-authored-by: Drew Powers <drew@pow.rs>